requirements.txt
pandas
numpy
sqlalchemy
psycopg2-binary
requests
fastapi
uvicorn[standard]
prefect
python-dotenv
pytest
httpx
pytest-asyncio
slowapi